Internal calculation of pixel values are done as double instead of int to avoid consequential errors when calculating (#1153)

+++ Tests/FO-DICOM.Tests/Imaging/LUT/PrecalculatedLUTTest.cs
@@ -40,10 +40,10 @@ namespace FellowOakDicom.Tests.Imaging.LUT
         }
 
         public bool IsValid { get; private set; }
-        public int MinimumOutputValue { get; private set; }
-        public int MaximumOutputValue { get; private set; }
+        public double MinimumOutputValue { get; private set; }
+        public double MaximumOutputValue { get; private set; }
 
-        public int this[int input] => input;
+        public double this[double input] => input;
 
         public void Recalculate()
         {
